Move contents of array.h into classlabels.h
